@@ -9,21 +9,26 @@ use WernerDweight\ApiAuthBundle\Entity\ApiClientInterface;
 
 class ApiClientCredentialsCheckedEvent extends Event
 {
-    /** @var string */
+    /**
+     * @var string
+     */
     public const NAME = 'wds.api_auth_bundle.api_client_credentials_checked';
 
-    /** @var bool */
+    /**
+     * @var bool
+     */
     private $valid = true;
 
-    /** @var ApiClientCredentials */
+    /**
+     * @var ApiClientCredentials
+     */
     private $credentials;
 
-    /** @var ApiClientInterface */
+    /**
+     * @var ApiClientInterface
+     */
     private $apiClient;
 
-    /**
-     * ApiClientCredentialsCheckedEvent constructor.
-     */
     public function __construct(ApiClientCredentials $credentials, ApiClientInterface $apiClient)
     {
         $this->credentials = $credentials;
@@ -40,13 +45,9 @@ class ApiClientCredentialsCheckedEvent extends Event
         return $this->apiClient;
     }
 
-    /**
-     * @return ApiClientCredentialsCheckedEvent
-     */
-    public function setValid(bool $valid): self
+    public function setValid(bool $valid): void
     {
         $this->valid = $valid;
-        return $this;
     }
 
     public function isValid(): bool
