@@ -7,7 +7,7 @@ use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
+use Symfony\Component\Security\Core\Exception\UserNotFoundException;
 use Symfony\Component\Security\Core\User\UserInterface;
 use WernerDweight\ApiAuthBundle\DTO\ApiUserCredentials;
 use WernerDweight\ApiAuthBundle\Entity\ApiUserInterface;
@@ -17,21 +17,34 @@ use WernerDweight\ApiAuthBundle\Service\ConfigurationProvider;
 
 class ApiUserLoader
 {
-    /** @var string */
+    /**
+     * @var string
+     */
     private const EXCEPTION_NOT_FOUND = 'There is no ApiUser for given api token!';
-    /** @var string */
+
+    /**
+     * @var string
+     */
     private const EXCEPTION_NO_SUCH_CREDENTIALS = 'There is no ApiUser for given credentials!';
 
-    /** @var EntityManager */
+    /**
+     * @var EntityManager
+     */
     private $entityManaager;
 
-    /** @var ConfigurationProvider */
+    /**
+     * @var ConfigurationProvider
+     */
     private $configurationProvider;
 
-    /** @var UserPasswordEncoderInterface */
+    /**
+     * @var UserPasswordEncoderInterface
+     */
     private $passwordEncoder;
 
-    /** @var ApiUserClassResolver */
+    /**
+     * @var ApiUserClassResolver
+     */
     private $apiUserClassResolver;
 
     /**
@@ -51,13 +64,6 @@ class ApiUserLoader
         $this->apiUserClassResolver = $apiUserClassResolver;
     }
 
-    private function getRepository(): ApiUserRepositoryInterface
-    {
-        /** @var ApiUserRepositoryInterface $repository */
-        $repository = $this->entityManaager->getRepository($this->apiUserClassResolver->getUserClass());
-        return $repository;
-    }
-
     /**
      * @throws \Safe\Exceptions\StringsException
      */
@@ -68,7 +74,7 @@ class ApiUserLoader
         /** @var ApiUserInterface|null $apiUser */
         $apiUser = $repository->findOneByToken($username);
         if (null === $apiUser) {
-            throw new UsernameNotFoundException(self::EXCEPTION_NOT_FOUND);
+            throw new UserNotFoundException(self::EXCEPTION_NOT_FOUND);
         }
         return $apiUser;
     }
@@ -80,14 +86,22 @@ class ApiUserLoader
     {
         $loginProperty = $this->configurationProvider->getUserLoginProperty();
         /** @var (ApiUserInterface & UserInterface)|null $user */
-        $user = $this->getRepository()->findOneBy([
-            $loginProperty => $credentials->getLogin(),
-        ]);
+        $user = $this->getRepository()
+            ->findOneBy([
+                $loginProperty => $credentials->getLogin(),
+            ]);
 
         if (null === $user || true !== $this->passwordEncoder->isPasswordValid($user, $credentials->getPassword())) {
             throw new UnauthorizedHttpException(ApiAuthEnum::REALM, self::EXCEPTION_NO_SUCH_CREDENTIALS);
         }
 
         return $user;
+    }
+
+    private function getRepository(): ApiUserRepositoryInterface
+    {
+        /** @var ApiUserRepositoryInterface $repository */
+        $repository = $this->entityManaager->getRepository($this->apiUserClassResolver->getUserClass());
+        return $repository;
     }
 }
