@@ -6,7 +6,7 @@ namespace WernerDweight\ApiAuthBundle\Security;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
-use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
+use Symfony\Component\Security\Core\Exception\UserNotFoundException;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
 use WernerDweight\ApiAuthBundle\Entity\ApiUserInterface;
@@ -15,16 +15,25 @@ use WernerDweight\ApiAuthBundle\Repository\ApiUserRepositoryInterface;
 
 class ApiUserRefresher
 {
-    /** @var string */
+    /**
+     * @var string
+     */
     private const EXCEPTION_NOT_FOUND = 'There is no ApiUser for given api token!';
-    /** @var string */
+
+    /**
+     * @var string
+     */
     private const EXCEPTION_UNSUPPORTED_USER =
         '%s is not a supported authentication class. Make sure your class implements ApiUserInterface!';
 
-    /** @var EntityManager */
+    /**
+     * @var EntityManager
+     */
     private $entityManaager;
 
-    /** @var ApiUserClassResolver */
+    /**
+     * @var ApiUserClassResolver
+     */
     private $apiUserClassResolver;
 
     /**
@@ -40,19 +49,12 @@ class ApiUserRefresher
         $this->apiUserClassResolver = $apiUserClassResolver;
     }
 
-    private function getRepository(): ApiUserRepositoryInterface
-    {
-        /** @var ApiUserRepositoryInterface $repository */
-        $repository = $this->entityManaager->getRepository($this->apiUserClassResolver->getUserClass());
-        return $repository;
-    }
-
     /**
      * @throws \Safe\Exceptions\StringsException
      */
     public function refresh(UserInterface $user): ApiUserInterface
     {
-        if (!$user instanceof ApiUserInterface) {
+        if (! $user instanceof ApiUserInterface) {
             throw new UnsupportedUserException(\Safe\sprintf(self::EXCEPTION_UNSUPPORTED_USER, get_class($user)));
         }
 
@@ -63,18 +65,24 @@ class ApiUserRefresher
             return $user;
         }
 
-        $id = $this->entityManaager
-            ->getClassMetadata($this->apiUserClassResolver->getUserClass())
-            ->getIdentifierValues($user);
-        if (true === empty($id)) {
+        $metadata = $this->entityManaager->getClassMetadata($this->apiUserClassResolver->getUserClass());
+        $id = $metadata->getIdentifierValues($user);
+        if (0 === count($id)) {
             throw new ApiUserProviderException(ApiUserProviderException::EXCEPTION_NO_ID);
         }
 
         /** @var ApiUserInterface|null $apiUser */
         $apiUser = $repository->find($id);
         if (null === $apiUser) {
-            throw new UsernameNotFoundException(self::EXCEPTION_NOT_FOUND);
+            throw new UserNotFoundException(self::EXCEPTION_NOT_FOUND);
         }
         return $apiUser;
+    }
+
+    private function getRepository(): ApiUserRepositoryInterface
+    {
+        /** @var ApiUserRepositoryInterface $repository */
+        $repository = $this->entityManaager->getRepository($this->apiUserClassResolver->getUserClass());
+        return $repository;
     }
 }
